# src/docs/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Hello Framework API
  version: 1.0.0
  description: Hello Framework REST API
  license:
    name: MIT

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      tags:
        - System
      responses:
        '200':
          description: Server is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                required:
                  - status

  /hello:
    get:
      summary: Get latest Hello message
      operationId: getLatestMessage
      tags:
        - Hello
      responses:
        '200':
          description: Hello message returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetHelloResponse'
        '404':
          description: Hello message not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create Hello message
      operationId: createMessage
      tags:
        - Hello
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHelloRequest'
      responses:
        '201':
          description: Hello message created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateHelloResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /hello/all:
    get:
      summary: Get all Hello messages
      operationId: getAllMessages
      tags:
        - Hello
      responses:
        '200':
          description: Hello messages returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetHelloResponse'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /hello/categories:
    get:
      summary: Get Hello categories
      operationId: getHelloCategories
      tags:
        - Hello
      responses:
        '200':
          description: Hello categories returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HelloCategoryResponse'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /hello/categories/{id}/messages:
    get:
      summary: Get Hello category with messages
      operationId: getHelloCategoryMessages
      tags:
        - Hello
      parameters:
        - name: id
          in: path
          required: true
          description: Hello category ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Hello category and messages returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelloCategoryWithMessagesResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Hello category not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /hello/{id}:
    get:
      summary: Get Hello message by ID
      operationId: getMessageById
      tags:
        - Hello
      parameters:
        - name: id
          in: path
          required: true
          description: Hello message ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Hello message returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetHelloResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Hello message not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    HelloCategoryResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: system
        description:
          type: string
          nullable: true
          example: Framework and system messages
      required:
        - id
        - name

    GetHelloResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        message:
          type: string
          example: Hello World
        category:
          nullable: true
          $ref: '#/components/schemas/HelloCategoryResponse'
      required:
        - id
        - message

    CreateHelloRequest:
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 255
          example: Hello World
        categoryId:
          type: integer
          example: 1
      required:
        - message

    CreateHelloResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        message:
          type: string
          example: Hello World
        category:
          nullable: true
          $ref: '#/components/schemas/HelloCategoryResponse'
      required:
        - id
        - message

    HelloCategoryWithMessagesResponse:
      allOf:
        - $ref: '#/components/schemas/HelloCategoryResponse'
        - type: object
          properties:
            messages:
              type: array
              items:
                $ref: '#/components/schemas/GetHelloResponse'
          required:
            - messages

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: No hello message found
      required:
        - error

    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: object
      required:
        - error